added the profile

+++ src/pages/Single.jsx
@@ -1,22 +1,22 @@
-import React, { useEffect, useState } from "react";
-import Edit from "../img/edit.png";
-import Delete from "../img/delete.png";
+import React, { useEffect, useState, useContext } from "react";
 import { Link, useLocation, useNavigate } from "react-router-dom";
-import Menu from "../components/Menu";
-import axios from "axios";
-import moment from "moment";
-import { useContext } from "react";
+
 import { AuthContext } from "../context/authContext";
 import DOMPurify from "dompurify";
 
+import axios from "axios";
+import moment from "moment";
+
+import Edit from "../img/edit.png";
+import Delete from "../img/delete.png";
+import Menu from "../components/Menu";
+
+
 const Single = () => {
     const [post, setPost] = useState({});
-
     const location = useLocation();
     const navigate = useNavigate();
-
     const postId = location.pathname.split("/")[2];
-
     const { currentUser } = useContext(AuthContext);
 
     useEffect(() => {
@@ -32,35 +32,29 @@ const Single = () => {
     }, [postId]);
 
     const handleDelete = async () => {
-        try {
-            const res = await axios.delete(`http://localhost:8000/v1/api/posts/delete/${postId}`,
-                {
+        if (window.confirm("do you want to delete this post")) {
+            try {
+                const res = await axios.delete(`http://localhost:8000/v1/api/posts/delete/${postId}`, {
                     headers: {
                         Authorization: `Bearer ${currentUser?.token}`,
                     },
                 });
-            if (res.status === 200) {
-                navigate("/")
-                return;
+                if (res.status === 200) {
+                    navigate("/");
+                    return;
+                }
+            } catch (err) {
+                console.log(err);
             }
-        } catch (err) {
-            console.log(err);
         }
-    }
+    };
 
-    const getText = (html) => {
-        const doc = new DOMParser().parseFromString(html, "text/html")
-        return doc.body.textContent
-    }
     return (
         <div className="single">
             <div className="content">
-                <img src={post?.image} alt="" />
+                {post.image && <img src={post.image} alt="" />}
                 <div className="user">
-                    {post.userImg && <img
-                        src={post.userImg}
-                        alt=""
-                    />}
+                    {post.userImg && <img src={post.userImg} alt="" />}
                     <div className="info">
                         <span>{post?.username}</span>
                         <p>Posted {moment(post?.date).fromNow()}</p>
@@ -75,11 +69,11 @@ const Single = () => {
                     )}
                 </div>
                 <h1>{post.title}</h1>
-                <p
-                    dangerouslySetInnerHTML={{
-                        __html: DOMPurify.sanitize(post.desc),
-                    }}
-                ></p>      </div>
+                <div
+                    dangerouslySetInnerHTML={{ __html: DOMPurify.sanitize(post?.desc || '') }}
+                    className="desc"
+                />
+            </div>
             <Menu category={post?.category} />
         </div>
     );
